Added the forEach() method.
It’s now possible to iterate over the elements of a mapping. Yay! I
made an intentional decision to not allow breaking out of forEach
loops… it seems like that could lead to a lot of bad code. I mean,
forEach means “do something for EVERY element in the object”, so if
nothing else it would be unsemantic.